+++ storage/ListByteBuffer.h
@@ -9,14 +9,24 @@
 #include "ListBufferByteTag.h"
 #include "ListElementView.h"
 
-#include "metadata/PropertyType.h"
-
 namespace db {
 
+/**
+ * @brief Stores list elements as a tag, followed by the raw bytes of the value of the
+ * element ("tag" in the sense of a "tagged union": storing type information). References
+ * are permanently stable.
+ *
+ * @detail Reference stability is guaranteed via storing elements in a linked list of
+ * contiguous chunks.
+ *
+ * @tparam N size (in bytes) of each chunk.
+ */
 template <size_t N = 4096>
 class ListByteBuffer {
 public:
+    /// Allocates the first chunk
     ListByteBuffer();
+    /// Deallocates all owned chunks
     ~ListByteBuffer();
 
     static consteval size_t tagSize() { return _tagSize; }
@@ -31,6 +41,8 @@ public:
     /**
      * @brief Writes the provided @param tag, followed by the bytes of @param val, in the
      * @ref _last byte buffer.
+     * @warn Must only be called after an appropriate call to @ref reserveContiguous.
+     * @warn Does not perform bounds checking.
      */
     template <typename T>
     ListElementView write(ListBufferTypeTag tag, const T& val);
@@ -38,58 +50,38 @@ public:
 private:
     class ByteChunk;
 
+    /// Head, Tail pointers of the linked list of chunks
     ByteChunk* _first {nullptr};
     ByteChunk* _last {nullptr};
 
+    /// Allocates a new chunk, updating @ref _last
     ByteChunk* allocateNextChunk();
 
     static constexpr size_t _tagSize = sizeof(ListBufferTypeTag);
 
-    static_assert(_tagSize == 1);
+    static_assert(_tagSize == 1, "Tag size changed");
     static_assert(N != 0);
     static_assert(N < std::numeric_limits<int64_t>::max());
 };
 
+/**
+ * @brief Internal class to manage the chunks of bytes of the outer @ref ListByteBuffer.
+ *
+ * @tparam N size (in bytes) of this chunk's buffer.
+ */
 template <size_t N>
 class ListByteBuffer<N>::ByteChunk {
 public:
     friend ListByteBuffer;
 
+    /// Returns true if @ref _buf has at least @param numBytes remaining capacity
     [[nodiscard]] bool canFit(size_t numBytes) const;
 
 private:
     std::array<std::byte, N> _buf;
+    /// The number of elements of @ref _buf that have been written to
     size_t _size {0};
     ByteChunk* _next {nullptr};
-};
-
-/// Helpers to convert types to tags
-template <typename T>
-struct TypeToListBufferTag;
-
-template <>
-struct TypeToListBufferTag<types::Int64::Primitive> {
-    static constexpr ListBufferTypeTag Tag = ListBufferTypeTag::Int;
-};
-
-template <>
-struct TypeToListBufferTag<types::UInt64::Primitive> {
-    static constexpr ListBufferTypeTag Tag = ListBufferTypeTag::UInt;
-};
-
-template <>
-struct TypeToListBufferTag<types::Double::Primitive> {
-    static constexpr ListBufferTypeTag Tag = ListBufferTypeTag::Double;
-};
-
-template <>
-struct TypeToListBufferTag<types::Bool::Primitive> {
-    static constexpr ListBufferTypeTag Tag = ListBufferTypeTag::Bool;
-};
-
-template <>
-struct TypeToListBufferTag<types::String::Primitive> {
-    static constexpr ListBufferTypeTag Tag = ListBufferTypeTag::String;
 };
 
 }
